refactor: useCountries, usePosts를 useCompany에 통합
