--- src/hooks/useCountries.ts
@@ -1,9 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-import { fetchCountries } from '@/lib/api';
-
-export function useCountries() {
-  return useQuery({
-    queryKey: ['countries'],
-    queryFn: fetchCountries,
-  });
-}
